==> obesity_predictions/01-initial-analysis.Rmd <==

# Initial Analysis

## R Packages and Setup {-}

```{r, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(janitor)
library(dplyr)
library(ggplot2)
library(corrplot)
library(skimr)
```

## Load Data

Let's begin by loading the dataset and examining its crucial variables and values.

```{r}
# Load data
raw <- read.csv('../obesity_level.csv')

# Display statistics
str(raw)
summary(raw)
head(raw)
```

According to the dataset, the following facts can be observed:

  - **Biometric Variables**:
    - **Gender (Gender)**: Indicates whether the individual is male or female.
    - **Age (Age)**: Specifies the age of each individual.
    - **Height (Height)**: Records the height in meters of each person in the dataset. (not used in our analysis)
    - **Weight (Weight)**: Captures the weight in kilograms of each individual. (not used in our analysis)
  - **Medical and Health-related Factors**:
    - **Family History with Overweight (family_history_with_overweight)**: Tracks whether individuals have a familial history of being overweight.
    - **Smoking Status(SMOKE)**: Indicates whether individuals are smokers or non-smokers.
  - **Dietary Habits**:
      - **Frequent Consumption of High-caloric Food (FAVC)**: Observes if individuals frequently consume high-caloric foods.
      - **Frequency of Consumption of Vegetables (FCVC)**: Measures how frequently individuals incorporate vegetables into their diet.
      - **Number of Main Meals (NCP)**: Records the number of main meals individuals consume in a day.
      - **Consumption of Food Between Meals (CAEC)**: Tracks individuals' consumption of food between regular meals.
      - **Daily Water Consumption (CH2O)**: Monitors the amount of water individuals consume daily in liters.
      - **Caloric Beverages Consumption (SCC)**: Notes the intake of caloric beverages.
      - **Consumption of Alcohol (CALC)**: Observes whether an individual consumes alcohol or not.
    - **Physical Activity**:
      - **Physical Activity Frequency (FAF)**: Records how often individuals engage in physical activities in a week.
      - **Time Spent Using Technological Devices (TUE)**: Measures the amount of time individuals spend using technological devices.
      - **Mode of Transportation (MTRANS)**: Observes the mode of transportation individuals use in their daily routines.
  - **Outcome Variable**:
    - **Obesity Level (0be1dad)**: indicates if someone is obese or not.


## Clean Data

We will perform data cleaning in the following steps, as the dataset is rather messy.

```{r}
# Copy raw data
cleaned <- raw

# Drop unnecessary columns
cleaned <- cleaned %>%
  select(-id, -Weight, -Height)

# Rename columns
cols_to_rename <- c(
  'Gender' = 'gender',
  'Age' = 'age',
  # 'Height' = 'height',
  # 'Weight' = 'weight',
  'family_history_with_overweight' = 'fam_history_overweight',
  'FAVC' = 'freq_consumption_hc_food',
  'FCVC' = 'veg_freq',
  'NCP' = 'num_meals',
  'CAEC' = 'food_between_meals',
  'SMOKE' = 'smoke',
  'CH2O' = 'water_daily',
  'SCC' = 'caloric_beverages',
  'FAF' = 'activity_freq',
  'TUE' = 'tech_use',
  'CALC' = 'alcohol',
  'MTRANS' = 'transport',
  'X0be1dad' = 'obesity_level'
)
names(cleaned) <- cols_to_rename[names(cleaned)]

# Simplify columns
cleaned <- cleaned %>%
  mutate(age = as.integer(round(age)),
         # height = round(height, digits = 2),
         # weight = round(weight, digits = 2),
         food_between_meals = ifelse(food_between_meals == 0,
                                     'Never', food_between_meals),
         alcohol = ifelse(alcohol == 0, 'Never', alcohol),
         obesity_level = case_when(
           obesity_level %in% c('Insufficient_Weight',
                                '0rmal_Weight') ~ 'normal_weight',
           obesity_level %in% c('Obesity_Type_I',
                                'Obesity_Type_II',
                                'Obesity_Type_III',
                                'Overweight_Level_I',
                                'Overweight_Level_II') ~ 'overweight',
           TRUE ~ obesity_level))

# Convert columns to factors
cols_to_factor <- c('gender',
                    'fam_history_overweight',
                    'freq_consumption_hc_food', 
                    'food_between_meals',
                    'smoke',
                    'alcohol',
                    'caloric_beverages',
                    'transport',
                    'obesity_level')
cleaned[cols_to_factor] <- lapply(cleaned[cols_to_factor], as.factor)

# Display data
str(cleaned)
summary(cleaned)
head(cleaned)
```


## Feature Engineering

With the dataset cleaned, we will now attempt some basic levels of feature engineering. This is beneficial when dealing with datasets having a limited number of columns, and can help reveal underlying dynamics between key variables. We will perform the following feature engineering techniques:

  - Create binary outcomes for eating habits based on individual records. Higher scores may suggest healthier eating habits.
  - Develop binary outcomes for physical activities.
  - Classify age groups categorically, which could help uncover age-specific trends related to obesity.

**Note that we are not using `BMI`, `Height`, or `Weight` as predictors given that we are trying to build models without these variables.**

```{r}
# Create BMI column
# cleaned$bmi <- cleaned$weight / (cleaned$height ^ 2)

# Create weight to height ratio
# cleaned$weight_to_height_ratio <- cleaned$weight / cleaned$height

# Create eating habit binary outcome
cleaned$eating_habit_binary <- (
  as.numeric(cleaned$freq_consumption_hc_food) +
    as.numeric(cleaned$veg_freq) +
    as.numeric(cleaned$num_meals))/3

# Create physical activity binary outcome
cleaned$physical_activity_binary <- cleaned$activity_freq - cleaned$tech_use

# Create age group
cleaned$age_group <- cut(cleaned$age,
                         breaks = c(-Inf, 18, 35, 50, Inf),
                         labels = c('Teen',
                                    'Young Adult',
                                    'Adult',
                                    'Senior'),
                         right = FALSE)

# Create water intake per weight
# cleaned$water_intake_per_weight <- (cleaned$water_daily / cleaned$weight) * 100

# Display data
str(cleaned)
summary(cleaned)
head(cleaned)
```


## Visualize Data

Now that all features have been added, we will proceed with our exploratory data analysis

```{r}
# Identify categorical and numerical columns
categorical_cols <- sapply(cleaned, is.factor)
numerical_cols <- sapply(cleaned, is.numeric)

# Exclude certain columns
exclude_cols <- c('weight', 'height')

# Frequency charts
for (col in names(cleaned)[categorical_cols]) {
  if (!(col %in% exclude_cols)) {
    freq_table <- table(cleaned[[col]])
    freq_df <- as.data.frame(freq_table)
    colnames(freq_df) <- c('Category', 'Frequency')
    plot <- ggplot(freq_df, aes(x = Category, y = Frequency)) +
      geom_bar(stat = 'identity') +
      labs(title = paste('Frequency Chart for', col)) +
      theme(axis.text.x = element_text(angle = 45, hjust = 1))
    print(plot)
  }
}

# Histograms
for (col in names(cleaned)[numerical_cols]) {
  if (!(col %in% exclude_cols)) {
    if (col == 'height') {
      plot <- ggplot(cleaned, aes(x = !!sym(col))) +
        geom_histogram(binwidth = 0.05, fill = 'skyblue', color = 'black') +
        labs(title = paste('Histogram for', col)) +
        theme(axis.text.x = element_text(angle = 45, hjust = 1))
    } else if (col == 'weight') {
      plot <- ggplot(cleaned, aes(x = !!sym(col))) +
        geom_histogram(binwidth = 10, fill = 'skyblue', color = 'black') +
        labs(title = paste('Histogram for', col)) +
        theme(axis.text.x = element_text(angle = 45, hjust = 1))
    } else {
      plot <- ggplot(cleaned, aes(x = !!sym(col))) +
        geom_histogram(binwidth = 1, fill = 'skyblue', color = 'black') +
        labs(title = paste('Histogram for', col)) +
        theme(axis.text.x = element_text(angle = 45, hjust = 1))
    }
    print(plot)
  }
}

# Correlation matrix
corrplot(cor(cleaned[, numerical_cols]), method="color", addCoef.col = "black", number.cex = 0.9, tl.cex = 1, tl.col="black", tl.srt=45, type = 'upper', order = 'alphabet', number.digits = 1)
```


Based on the visuals, we can see that:

  - Gender: This shows a nearly equal distribution between female and male categories, indicating that both genders are nearly equally represented in the dataset.
  - Family History of Overweight: There's a higher frequency for individuals with a family history of being overweight. This suggests that there is a potential genetic or familial component contributing to obesity, as many individuals in this sample have a family history of overweight conditions.
  - Frequency of High-Calorie Food Consumption: This shows that there is an overwhelmingly high frequency of individuals who consume high-calorie foods. Regular consumption of high-calorie food is strongly associated with weight gain and obesity.
  - Food Between Meals: This displays the frequency of eating between meals, with 'Sometimes' being the most common response. Frequent snacking can contribute to an increased caloric intake, which is a risk factor for obesity.
  - Smoking: This indicates that most individuals in the dataset do not smoke. While smoking has been associated with lower body weight due to nicotine's appetite suppressant effects, the high non-smoking frequency shown here indicates that, within this dataset, smoking status may not be a primary factor in obesity.
  - Caloric Beverage Consumption: A higher frequency of individuals in the dataset do not consume caloric beverages often. This suggests that habitual consumption of sugary drinks is not common in this sample. However, for those who do consume these beverages, it could be a contributing factor to caloric surplus and weight gain.
  - Alcohol Consumption: More individuals report 'Sometimes' consuming alcohol. Alcohol can be calorically dense and may contribute to weight gain if consumed in large or frequent amounts.
  - Age Group: 'Young Adult' is the most frequent age group in the dataset, followed by a significantly smaller number of 'Adults' and 'Teens,' with 'Seniors' being the least frequent. Age can influence metabolism, lifestyle choices, and thus risk of obesity.
  - Vegetable Consumption Frequency: This histogram may indicate how frequently individuals consume vegetables. A higher intake of vegetables is generally associated with healthier eating patterns, which can contribute to a lower risk of obesity.
  - Number of Meals: This shows the distribution of the number of meals participants report having each day. Eating patterns, including the number of meals daily, might influence weight and overall health, although the relationship can be complex.
  - Daily Water Intake: Proper hydration can affect appetite and metabolism, both of which play a role in maintaining a healthy weight.
  - BMI: BMI is a key index used to categorize individuals according to their risk of weight-related health problems. This histogram shows the BMI distribution, which directly relates to the study of obesity.
  - Weight to Height Ratio: This could be another indicator of body composition, similar to BMI but not necessarily the same calculation. Ratios that are higher might indicate higher levels of overweight or obesity.
  - Water Intake By weight: This chart shows the amount of water intake relative to body weight. Similar to daily water intake distribution chart, hydration status normalized by weight may provide insights into individual health behaviors.


## Split Data

We will initially split the data with a 50-50 ratio for train and test sets in our stacked models, reserving sufficient data for re-splitting during the second level model.

```{r}
# Copy cleaned data
split <- cleaned

# Dummy data
dummied <- as.data.frame(model.matrix(~.-1,split))

# Clean up data
dummied <- clean_names(dummied)

# Set seed for reproducibility
set.seed(12345)

# Define training set ratio
train_ratio <- 0.50

# Randomly sample training set
train_rows <- sample(1:nrow(dummied), train_ratio * nrow(dummied))
```


### Split Train Test

```{r}
# Train test split
train <- dummied[train_rows, ]
test <- dummied[-train_rows, ]
```

```{r, echo=FALSE}
# Write to csv
write.csv(train, '../Train Test Set/train.csv', row.names = FALSE)
write.csv(test, '../Train Test Set/test.csv', row.names = FALSE)
```


### Split LR

```{r}
# LR split
train_lr <- dummied[train_rows, ]
test_lr <- dummied[-train_rows, ]
```

```{r, echo=FALSE}
# Write to csv
write.csv(train_lr, '../Train Test Set/train_lr.csv', row.names = FALSE)
write.csv(test_lr, '../Train Test Set/test_lr.csv', row.names = FALSE)
```


### Split DT

```{r}
# DT Split
train_dt <- dummied[train_rows, ]
test_dt <- dummied[-train_rows, ]
```

```{r, echo=FALSE}
write.csv(train_dt, '../Train Test Set/train_dt.csv', row.names = FALSE)
write.csv(test_dt, '../Train Test Set/test_dt.csv', row.names = FALSE)
```


### Split SVM

```{r}
# DT Split
train_svm <- dummied[train_rows, ]
test_svm <- dummied[-train_rows, ]
```

```{r, echo=FALSE}
write.csv(train_svm, '../Train Test Set/train_svm.csv', row.names = FALSE)
write.csv(test_svm, '../Train Test Set/test_svm.csv', row.names = FALSE)
```


### Split RF

```{r}
# DT Split
train_rf <- dummied[train_rows, ]
test_rf <- dummied[-train_rows, ]
```

```{r, echo=FALSE}
write.csv(train_rf, '../Train Test Set/train_rf.csv', row.names = FALSE)
write.csv(test_rf, '../Train Test Set/test_rf.csv', row.names = FALSE)
```


### Split XGB

```{r}
# DT Split
train_xgb <- dummied[train_rows, ]
test_xgb <- dummied[-train_rows, ]
```

```{r, echo=FALSE}
write.csv(train_xgb, '../Train Test Set/train_xgb.csv', row.names = FALSE)
write.csv(test_xgb, '../Train Test Set/test_xgb.csv', row.names = FALSE)
```


### Split ANN

Remember that ANN necessitate normalized data. We will implement min-max normalization, excluding the outcome variable, for successful integration into our models.

```{r}
# ANN split

# Normalize data
normalize <- function(x) {
  return ((x - min(x)) / (max(x) - min(x)))
}

outcome_variable <- dummied$obesity_leveloverweight


dummied <- as.data.frame(lapply(
  dummied[,!names(dummied) %in% 'obesity_leveloverweight'], normalize))
dummied$obesity_leveloverweight <- outcome_variable

# ANN split
train_ann <- dummied[train_rows, ]
test_ann <- dummied[-train_rows, ]
```

```{r, echo=FALSE}
# Write to csv
write.csv(train_ann, '../Train Test Set/train_ann.csv', row.names = FALSE)
write.csv(test_ann, '../Train Test Set/test_ann.csv', row.names = FALSE)
```


### Split KNN

Similar to ANN, KNN also require a min-max normalized dataset for optimal performance. Additionally, it is essential to distinguish between predictor and outcome variables when running the model.

```{r}
# KNN split
train_knn <- train_ann
test_knn <- test_ann

train_knn_x <- train_knn
train_knn_x$obesity_leveloverweight <- NULL
train_knn_y <- train_knn$obesity_leveloverweight

test_knn_x <- test_knn
test_knn_x$obesity_leveloverweight <- NULL
test_knn_y <- test_knn$obesity_leveloverweight
```

```{r, echo=FALSE}
# Write to csv
write.csv(train_knn_y,'../Train Test Set/train_knn_y.csv', row.names = FALSE)
write.csv(train_knn_x,'../Train Test Set/train_knn_x.csv', row.names = FALSE)
write.csv(test_knn_y,'../Train Test Set/test_knn_y.csv', row.names = FALSE)
write.csv(test_knn_x,'../Train Test Set/test_knn_x.csv', row.names = FALSE)
```

Once all train and test sets have been split, we will proceed to module 1 for L1 modeling.
