==> obesity_predictions/03-artificial-neural-network.Rmd <==
# Artificial Neural Network (ANN)

## R Packages and Setup {-}


```{r, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(caret)
library(neuralnet)
```


## Introduction

Below paragraph is copied from the assignment, we are planning on updating this as we formulate our narrative for the final deliverable:

Artificial Neural Networks (ANNs) are modeled after the human brain and capable of learning and improving from experience. One strength is their ability to adapt to complex patterns and relationships in high-dimensional data, making them powerful tools for our dataset. However, ANNs require large amounts of training data and computational resources, which can make implementation time-consuming and resource-intensive.

We will now attempt using an ANN. Despite being considered a ‘black box’ model due to its difficulty in interpreting decision-making processes compared to models like DT, ANNs may perform slightly better than LR in terms of accuracy.


## Load Data

```{r}
# Load data
train_ann <- read.csv('../Train Test Set/train_ann.csv')
test_ann <- read.csv('../Train Test Set/test_ann.csv')

# Display statistics
str(train_ann)
summary(train_ann)
head(train_ann)
```


## Model ANN

### Simple ANN

We will initiate with a single neuron (single node) ANN to run a preliminary simple model. Once we have finished working on that model, we will create a more complex model consisting of one hidden layer and one neuron.

```{r, cache=TRUE}
# Build a model
model_ann <- neuralnet(obesity_leveloverweight ~ .,
                       data = train_ann,
                       hidden = 1)
```

```{r}
# Display summary
summary(model_ann)
```

### Complex ANN

Adding hidden layers increases computational requirements, so this will take significantly longer to run compared to a simple ANN.


```{r, cache=TRUE}
# Display summary
model_ann_complex <- neuralnet(obesity_leveloverweight ~ .,
                               data = train_ann,
                               hidden = c(1, 1), stepmax = 1e+07)

summary(model_ann_complex)

```



## Evaluate ANN

### Simple ANN

```{r}
# Make a prediction
prediction_ann_probability <- predict(model_ann, test_ann)
prediction_ann <- ifelse(prediction_ann_probability >= 0.5, 1, 0)
summary(prediction_ann_probability)


# Perform confusion matrix
cm_ann <- confusionMatrix(as.factor(prediction_ann),
                          as.factor(test_ann$obesity_leveloverweight),
                          positive = '1')
cm_ann
```

Based on the confusion matrix, we observe that the model achieves an accuracy of ``r round(cm_ann$overall[1], 4)``, sensitivity of ``r round(cm_ann$byClass[1], 4)``, and a Kappa coefficient of ``r round(cm_ann$overall[2], 4)``. We will assess these results towards the conclusion of the assignment when we have completed stacked models.


```{r}
pred <- ROCR::prediction(prediction_ann_probability, test_ann$obesity_leveloverweight)
perf <- ROCR::performance(pred, measure = "tpr", x.measure = "fpr")
auc <- ROCR::performance(pred, measure="auc")


auc <- round(auc@y.values[[1]],3)
roc.data <- data.frame(fpr=unlist(perf@x.values),
                       tpr=unlist(perf@y.values),
                       model="GLM")

ggplot(roc.data, aes(x=fpr, ymin=0, ymax=tpr)) +
    geom_ribbon(alpha=0.2, fill = "blue") +
    geom_line(aes(y=tpr), col = "blue") +
    geom_abline(intercept = 0, slope = 1, lty = "dashed") +
    labs(title = paste0("Simple ANN ROC Curve w/ AUC=", auc)) +
    theme_bw()
```

```{r, echo=FALSE}
# write to csv
write.csv(prediction_ann_probability,
          file = '../Predictions/prediction_ann_probability.csv',
          row.names = FALSE)

# Extract confusion matrix
write.csv(cm_ann$table,
          file = '../Confusion Matrix/cm_ann.csv',
          row.names = FALSE)
```

### Complex ANN

```{r}
# Make a prediction
prediction_ann_complex_probability <- predict(model_ann_complex, test_ann)
prediction_ann_complex <- ifelse(prediction_ann_complex_probability >= 0.5, 1, 0)
summary(prediction_ann_complex_probability)

# Perform confusion matrix
cm_ann_complex <- confusionMatrix(as.factor(prediction_ann_complex),
                                  as.factor(test_ann$obesity_leveloverweight),
                                  positive = '1')
cm_ann_complex

```

Based on the confusion matrix, we observe that the model achieves an accuracy of ``r round(cm_ann_complex$overall[1], 4)``, sensitivity of ``r round(cm_ann_complex$byClass[1], 4)``, and a Kappa coefficient of ``r round(cm_ann_complex$overall[2], 4)``. We will assess these results towards the conclusion of the assignment when we have completed stacked models.

```{r}
pred <- ROCR::prediction(prediction_ann_complex_probability, test_ann$obesity_leveloverweight)
perf <- ROCR::performance(pred, measure = "tpr", x.measure = "fpr")
auc <- ROCR::performance(pred, measure="auc")


auc <- round(auc@y.values[[1]],3)
roc.data <- data.frame(fpr=unlist(perf@x.values),
                       tpr=unlist(perf@y.values),
                       model="GLM")

ggplot(roc.data, aes(x=fpr, ymin=0, ymax=tpr)) +
    geom_ribbon(alpha=0.2, fill = "blue") +
    geom_line(aes(y=tpr), col = "blue") +
    geom_abline(intercept = 0, slope = 1, lty = "dashed") +
    labs(title = paste0("Complex ANN ROC Curve w/ AUC=", auc)) +
    theme_bw()
```

```{r, echo=FALSE}
# write to csv
write.csv(prediction_ann_complex_probability,
          file = '../Predictions/prediction_ann_complex_probability.csv',
          row.names = FALSE)

# Extract confusion matrix
write.csv(cm_ann_complex$table,
          file = '../Confusion Matrix/cm_ann_complex.csv',
          row.names = FALSE)
```
